refactor(settings): modularize SettingsForm into focused components
Break down monolithic 1200+ line SettingsForm.vue into 8 specialized components:

**Component Architecture:**
- Main SettingsForm.vue: Orchestration layer (~250 lines)
  * State management (iSettings, originalSettings)
  * Change tracking with deep watcher
  * API calls (load/submit settings)
  * Event coordination between child components

- src/components/settings/:
  * PresetSelector.vue - Preset buttons with detailed info dialog
  * SettingsBanners.vue - Unsaved/reindexing warning banners
  * SearchTab.vue - displayedAttributes, searchableAttributes, filterableAttributes, sortableAttributes
  * RelevancyTab.vue - rankingRules, distinctAttribute, stopWords, synonyms, typoTolerance
  * SynonymsSection.vue - Complete synonym CRUD with CSV import/export
  * TypoToleranceSection.vue - Typo tolerance configuration
  * PerformanceTab.vue - pagination, faceting with performance indicators
  * ReorderDialogs.vue - Drag-and-drop reorder dialogs for searchable/ranking

**Technical Details:**
- Props flow down: modelValue, hasFieldChanged function
- Events bubble up: apply-preset, show-searchable-reorder, show-ranking-reorder
- Fixed v-model on props error in ReorderDialogs (use :model-value + @update:model-value)
- Consistent orange border styling for modified fields
- Centralized notifications utility usage throughout

**Benefits:**
- Improved maintainability (focused components vs monolithic)
- Better code organization (single responsibility per component)
- Easier testing and debugging
- Reduced cognitive load when making changes

Files:
- Modified: src/components/SettingsForm.vue (1200→250 lines)
- Created: src/components/settings/PresetSelector.vue
- Created: src/components/settings/SettingsBanners.vue
- Created: src/components/settings/SearchTab.vue
- Created: src/components/settings/RelevancyTab.vue
- Created: src/components/settings/SynonymsSection.vue
- Created: src/components/settings/TypoToleranceSection.vue
- Created: src/components/settings/PerformanceTab.vue
- Created: src/components/settings/ReorderDialogs.vue
- Backup: src/components/SettingsForm.old.vue (original preserved)
